+++ urlextract.py
@@ -13,10 +13,12 @@ import re
 import string
 import sys
 import logging
+import tempfile
 import urllib.request
 import warnings
 from datetime import datetime, timedelta
 from urllib.error import URLError, HTTPError
+from appdirs import user_cache_dir
 
 import idna
 import uritools
@@ -54,7 +56,8 @@ class URLExtract:
             print("Given text contains some URL")
     """
     # file name of cached list of TLDs downloaded from IANA
-    _CACHE_FILE_NAME = '.urlextract_tlds'
+    _CACHE_FILE_NAME = 'tlds-alpha-by-domain.txt'
+    _DATA_DIR = 'data'
 
     # compiled regexp for naive validation of host name
     _hostname_re = re.compile(
@@ -71,43 +74,36 @@ class URLExtract:
         ("`", "`"),
     }
 
+    # name used in appdir
+    _URLEXTRACT_NAME = "urlextract"
+
     def __init__(self, cache_dir=None):
         """
         Initialize function for URLExtract class.
-        Tries to get cached .tlds, if cached file does not exist it will try
-        to download new list from IANNA and save it to users home directory.
+        Tries to get cached TLDs, if cached file does not exist it will try
+        to download new list from IANA and save it to users home directory.
 
-        :param str cache_dir: base path for tld cache, defaults to os.path.dirname(__file__)
+        :param str cache_dir: base path for TLD cache, defaults to data dir
         :raises: CacheFileError when cached file is not readable for user
         """
         self._logger = logging.getLogger(__name__)
 
-        # default to the current directory for cache
-        if cache_dir is None:
-            cache_dir = os.path.dirname(__file__)
-
-        # get directory for cached file
-        dir_path = cache_dir
-        if not os.access(dir_path, os.W_OK):
-            # get path to home dir
-            dir_path = os.path.expanduser('~')
+        # True if user specified path to cache directory
+        self._user_defined_cache = bool(cache_dir)
+        self._default_cache_file = False
 
         # full path for cached file with list of TLDs
-        self._tld_list_path = os.path.join(dir_path, self._CACHE_FILE_NAME)
+        self._tld_list_path = self._get_cache_file_path(cache_dir)
         if not os.access(self._tld_list_path, os.F_OK):
-            if not self._download_tlds_list():
-                raise CacheFileError("Can not download list of TLDs.")
-
-        # check if cached file is readable
-        if not os.access(self._tld_list_path, os.R_OK):
-            raise CacheFileError("Cached file is not readable for current "
-                                 "user. ({})".format(self._tld_list_path))
-
-        # try to update cache file when cache is older than 7 days
-        if not self.update_when_older(7):
-            self._logger.warning(
-                "Could not update file, using old version "
-                "of TLDs list. ({})".format(self._tld_list_path))
+            self._logger.info(
+                "Cache file not found in '%s'. "
+                "Use URLExtract.update() to download newest version.",
+                self._tld_list_path)
+            self._logger.info(
+                "Using default list of TLDs provided in urlextract package."
+            )
+            self._tld_list_path = self._get_default_cache_file_path()
+            self._default_cache_file = True
 
         self._tlds_re = None
         self._reload_tlds_from_file()
@@ -128,6 +124,84 @@ class URLExtract:
         # characters that are allowed to be right after TLD
         self._after_tld_chars = self._get_after_tld_chars()
 
+    def _get_default_cache_dir(self):
+        """
+        Returns default cache directory (data directory)
+
+        :raises: CacheFileError when default cached file does not is exist
+        :return: path to default cache directory
+        :rtype: str
+        """
+
+        return os.path.join(os.path.dirname(__file__), self._DATA_DIR)
+
+    def _get_default_cache_file_path(self):
+        """
+        Returns default cache file path
+
+        :return: default cache file path (to data directory)
+        :rtype: str
+        """
+
+        default_list_path = os.path.join(
+            self._get_default_cache_dir(), self._CACHE_FILE_NAME)
+
+        if not os.access(default_list_path, os.F_OK):
+            raise CacheFileError(
+                "Default cache file does not exist "
+                "'{}'!".format(default_list_path)
+            )
+
+        return default_list_path
+
+    def _get_writable_cache_dir(self):
+        """
+        Get writable cache directory with fallback to user's cache directory
+        and global temp directory
+
+        :raises: CacheFileError when cached directory is not writable for user
+        :return: path to cache directory
+        :rtype: str
+        """
+        dir_path_data = self._get_default_cache_dir()
+
+        if os.access(dir_path_data, os.W_OK):
+            self._default_cache_file = True
+            return dir_path_data
+
+        dir_path_user = user_cache_dir(self._URLEXTRACT_NAME)
+        if not os.path.exists(dir_path_user):
+            os.makedirs(dir_path_user, exist_ok=True)
+
+        if os.access(dir_path_user, os.W_OK):
+            return dir_path_user
+
+        dir_path_temp = tempfile.gettempdir()
+        if os.access(dir_path_temp, os.W_OK):
+            return dir_path_temp
+
+        raise CacheFileError("Cache directories are not writable.")
+
+    def _get_cache_file_path(self, cache_dir=None):
+        """
+        Get path for cache file
+
+        :param str cache_dir: base path for TLD cache, defaults to data dir
+        :raises: CacheFileError when cached directory is not writable for user
+        :return: Full path to cached file with TLDs
+        :rtype: str
+        """
+        if cache_dir is None:
+            # Tries to get writable cache dir with fallback to users data dir
+            # and temp directory
+            cache_dir = self._get_writable_cache_dir()
+        else:
+            if not os.access(cache_dir, os.W_OK):
+                raise CacheFileError("None of cache directories is writable.")
+
+        # get directory for cached file
+        return os.path.join(cache_dir, self._CACHE_FILE_NAME)
+
     def _get_after_tld_chars(self):
         """
         Initialize after tld characters
@@ -145,11 +219,15 @@ class URLExtract:
     def _reload_tlds_from_file(self):
         """
         Reloads TLDs from file and compile regexp.
+        :raises: CacheFileError when cached file is not readable for user
         """
         # check if cached file is readable
         if not os.access(self._tld_list_path, os.R_OK):
             self._logger.error("Cached file is not readable for current "
                                "user. ({})".format(self._tld_list_path))
+            raise CacheFileError(
+                "Cached file is not readable for current user."
+            )
         else:
             tlds = sorted(self._load_cached_tlds(), key=len, reverse=True)
             re_escaped = [re.escape(str(tld)) for tld in tlds]
@@ -165,7 +243,17 @@ class URLExtract:
         """
         url_list = 'https://data.iana.org/TLD/tlds-alpha-by-domain.txt'
 
-        # check if we can write cache file
+        # Default cache file exist (set by _default_cache_file)
+        # and we want to write permission
+        if self._default_cache_file and \
+                not os.access(self._tld_list_path, os.W_OK):
+            self._logger.info("Default cache file is not writable.")
+            self._tld_list_path = self._get_cache_file_path()
+            self._logger.info(
+                "Changed path of cache file to: %s",
+                self._tld_list_path
+            )
+
         if os.access(self._tld_list_path, os.F_OK) and \
                 not os.access(self._tld_list_path, os.W_OK):
             self._logger.error("ERROR: Cache file is not writable for current "
@@ -235,7 +323,7 @@ class URLExtract:
         """
         Update TLD list cache file.
 
-        :return: True if update was successfull False otherwise
+        :return: True if update was successful False otherwise
         :rtype: bool
         """
 
@@ -249,16 +337,16 @@ class URLExtract:
     def update_when_older(self, days):
         """
         Update TLD list cache file if the list is older than
-        number of days given in parameter `days`.
+        number of days given in parameter `days` or if does not exist.
 
         :param int days: number of days from last change
-        :return: True if update was successfull, False otherwise
+        :return: True if update was successful, False otherwise
         :rtype: bool
         """
 
         last_cache = self._get_last_cachefile_modification()
         if last_cache is None:
-            return False
+            return self.update()
 
         time_to_update = last_cache + timedelta(days=days)
 
@@ -728,6 +816,7 @@ def _urlextract_cli():
 
     try:
         urlextract = URLExtract()
+        urlextract.update_when_older(30)
         content = args.input_file.read()
         for url in urlextract.find_urls(content, args.unique):
             print(url)